+++ tooling/module-domain/src/test/java/org/eclipse/dataspaceconnector/tooling/module/domain/ServiceReferenceTest.java
@@ -1,0 +1,35 @@
+/*
+ *  Copyright (c) 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.tooling.module.domain;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class ServiceReferenceTest {
+
+    @Test
+    void verifySerializeDeserialize() throws JsonProcessingException {
+        var mapper = new ObjectMapper();
+        var service = new ServiceReference("foo.bar.BarService", false);
+
+        var serialized = mapper.writeValueAsString(service);
+        var deserialized = mapper.readValue(serialized, ServiceReference.class);
+
+        assertThat(deserialized).isNotNull();
+    }
+}